// features/custom-order/types.ts
export type Usage = 'cuisine' | 'pliant' | 'outdoor' | 'chasse'
// ...
export interface SteelOption {
  id: string
  label: string
  description: string
  imageSrc?: string
  imageAlt?: string
  tech: {
    retention: number
    sharpening: number
    flexibility: number
    price: number
  }
}